change setValue/getValue APIs with uint64_t instead of uint32_t according to the latest p4lang/tdi.

+++ src/tdi_rt/tdi_common/tdi_rt_target.cpp
@@ -37,7 +37,7 @@ namespace pna {
 namespace rt {
 
 tdi_status_t Target::setValue(const tdi_target_e &target_field,
-                              const uint32_t &value) {
+                              const uint64_t &value) {
     tdi_status_t status = TDI_SUCCESS;
     switch(static_cast<pna_target_e>(target_field)) {
       case PNA_TARGET_PIPE_ID:
@@ -57,7 +57,7 @@ tdi_status_t Target::setValue(const tdi_target_e &target_field,
 }
 
 tdi_status_t Target::getValue(const tdi_target_e &target_field,
-                              uint32_t *value) const {
+                              uint64_t *value) const {
     tdi_status_t status = TDI_SUCCESS;
     switch(static_cast<pna_target_e>(target_field)) {
       case PNA_TARGET_PIPE_ID:
